+++ examples/drawing.rs
@@ -5,6 +5,8 @@ use ggez::{Context, GameResult};
 use ggez::graphics;
 use ggez::graphics::{DrawMode, Point2};
 use ggez::timer;
+use std::env;
+use std::path;
 
 struct MainState {
     image1: graphics::Image,
@@ -112,6 +114,15 @@ impl event::EventHandler for MainState {
 pub fn main() {
     let c = conf::Conf::new();
     let ctx = &mut Context::load_from_conf("drawing", "ggez", c).unwrap();
+
+    // We add the CARGO_MANIFEST_DIR/resources do the filesystems paths so 
+    // we we look in the cargo project for files.
+    if let Ok(manifest_dir) = env::var("CARGO_MANIFEST_DIR") {
+        let mut path = path::PathBuf::from(manifest_dir);
+        path.push("resources");
+        ctx.filesystem.mount(&path, true);
+    }
+    
     println!("{}", graphics::get_renderer_info(ctx).unwrap());
     let state = &mut MainState::new(ctx).unwrap();
     if let Err(e) = event::run(ctx, state) {