+++ examples/imageview.rs
@@ -1,7 +1,5 @@
 extern crate ggez;
 extern crate rand;
-extern crate sdl2;
-
 
 use ggez::audio;
 use ggez::conf;
@@ -10,6 +8,8 @@ use ggez::{Context, GameResult};
 use ggez::graphics;
 use ggez::graphics::Color;
 use ggez::timer;
+use std::env;
+use std::path;
 
 struct MainState {
     a: i32,
@@ -122,6 +122,15 @@ pub fn main() {
     let c = conf::Conf::new();
     println!("Starting with default config: {:#?}", c);
     let ctx = &mut Context::load_from_conf("imageview", "ggez", c).unwrap();
+
+    // We add the CARGO_MANIFEST_DIR/resources do the filesystems paths so 
+    // we we look in the cargo project for files.
+    if let Ok(manifest_dir) = env::var("CARGO_MANIFEST_DIR") {
+        let mut path = path::PathBuf::from(manifest_dir);
+        path.push("resources");
+        ctx.filesystem.mount(&path, true);
+    }
+    
     let state = &mut MainState::new(ctx).unwrap();
     if let Err(e) = event::run(ctx, state) {
         println!("Error encountered: {}", e);